chores/ if bable-cli is not installed globally  it will be installed

+++ install.sh
@@ -1,7 +1,13 @@
 #  This script should do all the preparation for your project to run, such as downloading any dependencies and compiling if necessary
 echo "Installing dependencies..."
 npm install
-npm install -g babel-cli
+if test -s /usr/bin/babel
+then
+  echo "babel-cli has already been installed globally!"
+else
+  echo "Installing babel-cli globally..."
+  npm install -g babel-cli
+fi
 wait
 echo "Install complete!"
 echo "Transpiling..."
